Let system choose gpg package name
Some systems prefer `gnupg2` to `gpg`, so leave it up to the system
maintainer. Apt won't work without some version of gpg, so we can assume
it's present. The `apt-transport-https` req must be explicitly declared,
however.

Minor updates to tests for sanity-checking during local dev.

N.B. you may need to edit the name of the role in
`molecule/default/playbook.yml` for Ansible to find the path.